llm-silver-anno/requirements.txt
mmif-python~=1.0
tqdm
clams-utils
opencv_python==4.9.0.80
pandas==2.2.2
streamlit==1.33.0
streamlit_extras==0.4.3
streamlit_shortcuts==0.1.1
